Create Sal's Shipping
1.
First off, we need to know how much it costs to ship a package of given weight by normal ground shipping. Write a function for the cost of ground shipping. This function should take one parameter, weight, and return the cost of shipping a package of that weight.
2.
A package that weighs 8.4 pounds should cost $53.60 to ship with normal ground shipping:
8.4\ lb \times \$4.00 + \$20.00 = \$53.608.4 lb×$4.00+$20.00=$53.60
Test that your ground shipping function gets the same value.

3.
We’ll also need to make sure we include the price of premium ground shipping in our code. Create a variable for the cost of premium ground shipping.
Note: this does not need to be a function because the price of premium ground shipping does not change with the weight of the package.

4.
Write a function for the cost of drone shipping. This function should take one parameter, weight, and return the cost of shipping a package of that weight.

5.
A package that weighs 1.5 pounds should cost $6.75 to ship by drone:
1.5\ lb \times \$4.50 + \$0.00 = \$6.751.5 lb×$4.50+$0.00=$6.75
Test that your drone shipping function gets the same value.

6.
Using those two functions for ground shipping cost and drone shipping cost, as well as the cost of premium ground shipping, write a function that takes one parameter, weight and prints a statement that tells the user
The shipping method that is cheapest.
How much it would cost to ship a package of said weight using this method.

7.
Great job! Now, test your function!
What is the cheapest method of shipping a 4.8 pound package and how much would it cost?
What is the cheapest method of shipping a 41.5 pound package and how much would it cost?